Read entries and parseTitle from text
This commit is being made to read the "entries" key in the dayone
journal JSON. "entries" is an array of "entry" objects which contain all
of the data regarding the entry for that day/time.

The title needs to be parsed from the "text" key, as it's a massive
string. The text is formatting with escaped \n\n symbols. By splitting
the journal into paragraphs, the title can be found by looking at the
first paragraph, or a truncation of the first paragraph.

In the case that there are no paragraphs, or the journal is empty, the
title will be set to "Untitled".