Parse and build meta and content objects
In this commit, entries are read from the source day one journal. The
entries are then parsed for information like the title, creation date,
etc.

There are two objects which will be used to construct meta.json and
content.json for the Quiver data format.

Link parsing and title parsing are currently being developed and tested.